Backend and more
Supabase backend was setup with tables created and calls to the database are also setup in the host and player pages for creating a game or submitting guesses. Just need to attach these calls to the database to actual buttons within the app.     Also the player page was just added and the host page was seperated out with redirecting now used to go between them. some very slight css changes made too.